requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
Django==5.0.6
djangorestframework==3.15.1
exceptiongroup==1.2.1
fake-useragent==1.5.1
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
joblib==1.4.2
lxml==5.2.1
lxml_html_clean==0.1.1
nltk==3.8.1
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.2
parse==1.20.1
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.10
requests==2.31.0
requests-html==0.10.0
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
zipp==3.18.1
